Created a new script.
By Calling install-scripts it calls a helper script which installs all
the programs specified in the programs array. This will be useful when I
need to reinstall an OS and have this reinstall everything for me.

it also times the execution time, and sends the output to a log file.