Added some functions.
There is now a toggle to toggle between relative line numbering, and non
relative line numbering.

My formatting and whitespace substitution auto commands are now all in a
function that saves the last known cursor position so that you do not lose your
place in the file after a save that makes changes.