Now auto installs more automatically.
This script should no longer ask for user inputs to install a program (such as
entering a y to install a program).

Also Added an email feature, that lets a command line argument of --email so
that the script will email you when the installation is finished with the log.
(note that mailutils must be installed for this feature to work).

I also added clamav to the install list, which is a virus scanner for linux.