I added a new program to my setup.
Courtesty of https://github.com/nvbn/thefuck a program called "thefuck" is now
installed.

The files can be found in ~/.thefuck

This program will fix your previous command if it was entered wrong.

I also added a bash alias for this program, and added a few programs to the list
of programs to install that need to be installed in order to install this
program.