Added a script to try and make ubuntu run faster.
Swapiness is reduced to a 10 so that the swap partition is not used as often.

Hidden startup programs are no longer hidden, and thus more programs can be
removed.

Also installs preload if not already installed, which preloads your most used
programs into memory on boot. This has also been added to the auto install
script.