Improved the K command.
K was used in my vimrc to move the current line up 1 line. It had a problem when
you tried to move a line that was the first or last line of the file though.

If it was the first line of a file it would delete the line altogether.

If it was the last line, it would move up 2 spaces instead of 1.

This commit fixes those problems.