Big Changes
The repo is now setup to be cloned into a separate directory from ~/

This I dont get information from all the subfolders of ~/ about the status of
the home repo.

Because of this, this updated repo is not going to work well with repos that
already have a previous clone.