A start towards making the install script work on Manjaro
Currently the script assumes you are using a distro with apt-get. I
recently started looking at Manjaro which is based off of Arch, which
means that the package manager now is pacman, not apt-get.

This change does not fix all the problems, but it at least allows some
programs to be installed now. There is no support currently for
updating, or for adding new repos, and some programs still fail to
install because the package names appear to be different.