Added the ability to toggle lightline themes
Now when toggling colorschemes, the lightline colour theme will also
change. The default is powerline, but if the 3rd field of the colours
list is a 1, it will use that colorschemes lightline theme.

I also made it so that when vim starts up, it calls the toggle function.
This way there is less duplicated code.