Filetype specific colourschemes are now enabled.
Add an entry to the new function GetDefaultColour for a filetype that
you want a specific colourscheme for.

In my case, the only entry as of the time of this commit is for lua
because I found that my default colourscheme is not great with lua
code. So my default lua colourscheme is now badwolf, but my default
scheme for everything else is gruvbox.