Treat long line wrapped lines as different lines for moving to end and beginning of lines
For the "H" and "L" maps, only move to the end and beginning of the line
as shown by the line wrap.

Also, When moving to the end of the line, move first to the closest
space character to the 80th column, so long as that space character is
the between the 60th and 81st column. If there are no spaces in that
span, go to the 80th column.

If I am already to the right of the space character I would have moved
to, or the 80th column, I would go to the end of the line.