Fixed my highlighting issues finally!!
I figured out that my problem was because colorschemes tend to clear
highlighting, which doesn't work so well if you want custom
highlighting, and if you want to be able to change colourschemes
regularly like I do.

So now, I added my highlights to a function that is automatically called
every time I change my colourscheme.