Installed gundo plugin
Turns out that vim saves your undo levels as a tree not a list which
means that you can type something, undo it, type something else, and
undo back to the original. The builtin way to use this in vim is not
very nice though, so this plugin makes that a bit easier and nicer to
use.