Added bullets.vim plugin.
This makes creating bullets or lists in vim much easier, as the bullets
or numbers are automatically filled in when I go to a new line.

I also have a custom setting for bullets.vim saying what types of files
to use the bullets.vim plugin with.

In order to get it to work with files with no filetype I also needed to
make a new autocmd that would change the filetype of a file with no
filetype to be "noft".