Fixed the mapping that corrects previous spelling
Before, the mapping wouldn't work if I was on the current word, but at
the end of the word while at the end of the line. This was a common use
case for me, as often I will finish writing a word, realize it is wrong,
and immediately go to fix it.

Another case that still doesn't work is if all the above is still true,
but I am also on the last line. A more complex function will probably be
needed to resolve that though and there probably isn't much point in
that.

I do not actually understand why the old mapping didn't work. It does
work when you do it manually. Doing an "l" movement must produce a delay
for vim in some way when it is on the end of the line that is messing
the mapping up.